+++ moleculer.config.ts
@@ -48,9 +48,9 @@ const brokerConfig: BrokerOptions = {
   // Retry policy settings. More info: https://moleculer.services/docs/0.13/fault-tolerance.html#Retry
   retryPolicy: {
     // Enable feature
-    enabled: true,
+    enabled: false,
     // Count of retries
-    retries: 5,
+    retries: 0,
     // First delay in milliseconds.
     delay: 100,
     // Maximum delay in milliseconds.