+++ services/products.service.ts
@@ -9,7 +9,7 @@ const TheService: ServiceSchema = {
   retryPolicy: {
     retries: 1,
   },
-  mixins: [I18nService, ProductTransformation, ESService, ProductsOpenapi, ProductsValidation, AppSearch('catalog')],
+  mixins: [I18nService, ProductTransformation, ESService, ProductsOpenapi, ProductsValidation, AppSearch(process.env.APP_SEARCH_ENGINE)],
   settings: {
     elasticsearch: {
       host: process.env.ELASTIC_URL,
@@ -236,7 +236,7 @@ const TheService: ServiceSchema = {
             [x: string]: string | number | string[];
           } = {
             id: product.id,
-            [product.attribute]: product.qty + 1,
+            [product.attribute]: product.qty,
           };
           if (product.imported) body.imported = product.imported;
           return body;