+++ utilities/mixins/openapi/subscription.openapi.mixin.ts
@@ -221,6 +221,7 @@ const SubscriptionCreateOpenapi = {
             membership: { type: 'string', required: true },
             coupon: { type: 'string' },
             grantTo: { type: 'string', format: 'url', description: 'This field is used to donor the subscription to another store' },
+            autoRenew: { type: 'boolean' },
           },
         },
       },