+++ utilities/mixins/validation/subscription.validate.ts
@@ -4,7 +4,7 @@ import { ServiceSchema } from 'moleculer';
 export const SubscriptionValidation: ServiceSchema = {
   name: 'subscription',
   actions: {
-    get: {
+    sGet: {
       params: {
         id: {
           type: 'string',
@@ -160,6 +160,10 @@ export const SubscriptionValidation: ServiceSchema = {
         },
         grantTo: {
           type: 'url',
+          optional: true,
+        },
+        autoRenew: {
+          type: 'boolean',
           optional: true,
         },
         $$strict: true,